+++ docs/CHANGES.rst
@@ -4,14 +4,19 @@ Changes
 =======
 
 
-1.11.0 (unreleased)
+1.11.1 (unreleased)
 -------------------
+- Add generic concentratedRunoffSRisk and details optional fields
+  [fngaha]
+
 - Add getFirstAdministrativeSfolderManager method
   [fngaha]
 
 - Add removeItems utils and listSolicitOpinionsTo method
   [fngaha]
 
+1.11.0 (2015-10-01)
+-------------------
 - Add getFirstDeposit and _getFirstEvent method
   [fngaha]